Move CLI methods to command_line_interface.py. main.py is now just the main functions asked for in the assignment prompt
