+++ setup.cfg
@@ -69,7 +69,7 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    todo = to_do_app.main:run
+    todo = to_do_app.command_line_interface:run
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
