+++ src/to_do_app/__init__.py
@@ -1,15 +1,10 @@
-import sys
-
-if sys.version_info[:2] >= (3, 8):
-    # TODO: Import directly (no need for conditional) when `python_requires = >= 3.8`
-    from importlib.metadata import PackageNotFoundError, version  # pragma: no cover
-else:
-    from importlib_metadata import PackageNotFoundError, version  # pragma: no cover
+""" to_do_app initialization file """
+from importlib.metadata import PackageNotFoundError, version  # pragma: no cover
 
 try:
     # Change here if project is renamed and does not equal the package name
-    dist_name = __name__
-    __version__ = version(dist_name)
+    DIST_NAME = __name__
+    __version__ = version(DIST_NAME)
 except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
 finally:
